Encapsulate MessageDigestPrototype in Sha256Hash
Avoid potential class shadowing conflicts by embedding
MessageDigestPrototype in Sha256Hash.
